@@ -20,15 +20,15 @@ classifiers =
     Programming Language :: Python
     Programming Language :: Python :: 3
     Programming Language :: Python :: 3 :: Only
-    Programming Language :: Python :: 3.5
     Programming Language :: Python :: 3.6
     Programming Language :: Python :: 3.7
     Programming Language :: Python :: 3.8
+    Programming Language :: Python :: 3.9
     Topic :: Software Development :: Libraries
     Topic :: Utilities
 
 [options]
-python_requires = >=3.5
+python_requires = >=3.6
 packages =
     django_redis
     django_redis.client
