@@ -5,7 +5,7 @@ class ConnectionInterrupted(Exception):
     def __str__(self):
         error_type = type(self.__cause__).__name__
         error_msg = str(self.__cause__)
-        return "Redis {}: {}".format(error_type, error_msg)
+        return f"Redis {error_type}: {error_msg}"
 
 
 class CompressorError(Exception):
