@@ -8,10 +8,10 @@ class Node:
         self.id = id
 
     def __str__(self):
-        return "node:{}".format(self.id)
+        return f"node:{self.id}"
 
     def __repr__(self):
-        return "<Node {}>".format(self.id)
+        return f"<Node {self.id}>"
 
 
 class HashRingTest(unittest.TestCase):
@@ -26,12 +26,12 @@ class HashRingTest(unittest.TestCase):
     def test_hashring(self):
         ids = []
 
-        for key in ["test{}".format(x) for x in range(10)]:
+        for key in [f"test{x}" for x in range(10)]:
             node = self.ring.get_node(key)
             ids.append(node.id)
 
         self.assertEqual(ids, [0, 2, 1, 2, 2, 2, 2, 0, 1, 1])
 
     def test_hashring_brute_force(self):
-        for key in ("test{}".format(x) for x in range(10000)):
+        for key in (f"test{x}" for x in range(10000)):
             self.ring.get_node(key)
